move twitter specific code into platform package
