+++ src/main/java/xyz/nedderhoff/citytweets/platform/twitter/TwitterHttpEndpoint.java
@@ -1,4 +1,4 @@
-package xyz.nedderhoff.citytweets.twitter;
+package xyz.nedderhoff.citytweets.platform.twitter;
 
 import org.springframework.web.client.RestTemplate;
