+++ src/main/java/xyz/nedderhoff/citytweets/platform/twitter/api1/FollowEndpoint.java
@@ -1,4 +1,4 @@
-package xyz.nedderhoff.citytweets.twitter.api1;
+package xyz.nedderhoff.citytweets.platform.twitter.api1;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,7 +10,7 @@ import xyz.nedderhoff.citytweets.cache.FriendCache;
 import xyz.nedderhoff.citytweets.cache.Twitter4jConnectionsCache;
 import xyz.nedderhoff.citytweets.config.AccountProperties.Account;
 import xyz.nedderhoff.citytweets.domain.User;
-import xyz.nedderhoff.citytweets.twitter.TwitterApi1Endpoint;
+import xyz.nedderhoff.citytweets.platform.twitter.TwitterApi1Endpoint;
 
 @Component
 public class FollowEndpoint extends TwitterApi1Endpoint {
