+++ src/main/java/xyz/nedderhoff/citytweets/platform/twitter/api1/FriendsEndpoint.java
@@ -1,4 +1,4 @@
-package xyz.nedderhoff.citytweets.twitter.api1;
+package xyz.nedderhoff.citytweets.platform.twitter.api1;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,7 +9,7 @@ import twitter4j.IDs;
 import twitter4j.TwitterException;
 import xyz.nedderhoff.citytweets.cache.Twitter4jConnectionsCache;
 import xyz.nedderhoff.citytweets.config.AccountProperties.Account;
-import xyz.nedderhoff.citytweets.twitter.TwitterApi1Endpoint;
+import xyz.nedderhoff.citytweets.platform.twitter.TwitterApi1Endpoint;
 
 import java.util.Arrays;
 import java.util.HashSet;
