+++ src/main/java/xyz/nedderhoff/citytweets/platform/twitter/api1/MeEndpoint.java
@@ -1,4 +1,4 @@
-package xyz.nedderhoff.citytweets.twitter.api1;
+package xyz.nedderhoff.citytweets.platform.twitter.api1;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,7 +8,7 @@ import org.springframework.web.client.RestTemplate;
 import twitter4j.TwitterException;
 import xyz.nedderhoff.citytweets.cache.Twitter4jConnectionsCache;
 import xyz.nedderhoff.citytweets.config.AccountProperties.Account;
-import xyz.nedderhoff.citytweets.twitter.TwitterApi1Endpoint;
+import xyz.nedderhoff.citytweets.platform.twitter.TwitterApi1Endpoint;
 
 @Component
 public class MeEndpoint extends TwitterApi1Endpoint {
